updated main.py with functions from robin and kaio
